requirements.txt
django==5.1.1
djangorestframework==3.15.2
pytest==8.3.3
pytest-django==4.9.0
pytest-mock
psycopg2-binary
pandas
gunicorn
dj_database_url
drf-yasg